time for some Christmas presents... libw32 compiles here without a single compiler warning
CVS patchset: 5946
CVS date: 2003/12/24 16:55:35

+++ src/libw32dll/DirectShow/DS_Filter.c
@@ -236,7 +236,7 @@ DS_Filter* DS_FilterCreate(const char* dllname, const GUID* id,
     if (!init)
     {
 	DS_Filter_Destroy(This);
-	printf("Warning: DS_Filter() %s.  (DLL=%.200s, r=0x%x)\n", em, dllname, result);
+	printf("Warning: DS_Filter() %s.  (DLL=%.200s, r=0x%x)\n", em, dllname, (unsigned)result);
         This = 0;
     }
     return This;
